Support for sortened command names
This feature exploits the fact that CLAP performs some simple heuristics
on the input it receives, and allows writing shortest uniquly
identifying command names instead of full names.
However, this only works if shortened command name is followed by a
string that CLAP recognizes as an option (a limitation of current
heuristics).